backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
django-cors-headers>=4.7,<4.8
dj-database-url>=2.3,<2.4
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.1,<1.2
